Moved LearningStandards property out of the OdsInstanceSettingsModel to have a LearninigStandardsIndexModel. Moved the LearningStandards viewmodels to the LearningStandards folder.
