+++ Application/EdFi.Ods.AdminApp.Web/Models/ViewModels/LearningStandards/LearningStandardsModel.cs
@@ -1,4 +1,4 @@
-﻿// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: Apache-2.0
 // Licensed to the Ed-Fi Alliance under one or more agreements.
 // The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
 // See the LICENSE and NOTICES files in the project root for more information.
@@ -7,7 +7,7 @@ using System;
 using System.ComponentModel.DataAnnotations;
 using FluentValidation;
 
-namespace EdFi.Ods.AdminApp.Web.Models.ViewModels
+namespace EdFi.Ods.AdminApp.Web.Models.ViewModels.LearningStandards
 {
     public class LearningStandardsModel
     {
